+++ backend/canisters/user/impl/src/memory.rs
@@ -9,7 +9,7 @@ pub type Memory = VirtualMemory<DefaultMemoryImpl>;
 
 thread_local! {
     static MEMORY_MANAGER: MemoryManager<DefaultMemoryImpl>
-        = MemoryManager::init_with_bucket_size(DefaultMemoryImpl::default(), 1);
+        = MemoryManager::init_with_bucket_size(DefaultMemoryImpl::default(), 2);
 }
 
 pub fn get_upgrades_memory() -> Memory {